+++ packages/multichain-account-service/src/providers/EvmAccountProvider.ts
@@ -15,6 +15,7 @@ import {
   EthAccountType,
   EthScope,
 } from '@metamask/keyring-api';
+import type { KeyringCapabilities } from '@metamask/keyring-api/v2';
 import { KeyringTypes } from '@metamask/keyring-controller';
 import type {
   EthKeyring,